ML_Classifiers_Comparison/requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
